# docs/openapi.yml
openapi: "3.0.3"
info:
  title: bloop API
  description: |
    Self-hosted error observability and LLM tracing API.

    bloop ingests error events from mobile apps and APIs, deduplicates them via
    fingerprinting, and provides query endpoints for reviewing errors. Optional
    feature-gated modules add DuckDB-powered analytics and LLM call tracing.

    **Conventions**

    - Timestamps are Unix epoch **milliseconds** unless noted otherwise.
    - LLM costs are stored as **microdollars** (1 dollar = 1,000,000 micros).
    - All error responses use `{ "error": "message" }`.
    - Pagination defaults: `limit=50`, `offset=0`, max `limit=200`.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    hmac:
      type: apiKey
      in: header
      name: X-Bloop-Signature
      description: |
        HMAC-SHA256 signature computed over the raw request body using the
        project API key. Used by ingest endpoints.

    bearer:
      type: http
      scheme: bearer
      description: |
        Bearer token issued via the management API. Each token is scoped to a
        project and carries permission scopes such as `errors:read` and
        `errors:write`.

    session:
      type: apiKey
      in: cookie
      name: bloop_session
      description: |
        Cookie-based session set after WebAuthn passkey authentication.
        Used by dashboard and management endpoints.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "validation failed: message is required"

    # -- Ingest --
...
    IngestSpan:
      type: object
      properties:
        id:
          type: string
          nullable: true
          description: Auto-generated UUID if omitted.
        parent_span_id:
          type: string
          nullable: true
        span_type:
          type: string
          default: generation
          enum: [generation, tool, retrieval, custom]
        name:
          type: string
          default: ""
        model:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        input_tokens:
          type: integer
          format: int64
          default: 0
        output_tokens:
          type: integer
          format: int64
          default: 0
        cost:
          type: number
          format: double
          default: 0
          description: Cost in dollars. Converted to microdollars at ingest. If zero and model is known, server-side pricing is applied.
        latency_ms:
          type: integer
          format: int64
          default: 0
        time_to_first_token_ms:
          type: integer
          format: int64
          nullable: true
        status:
          type: string
          default: ok
        error_message:
          type: string
          nullable: true
        input:
          type: string
          nullable: true
          description: Subject to content storage policy.
        output:
          type: string
          nullable: true
          description: Subject to content storage policy.
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        started_at:
          type: integer
          format: int64
          nullable: true
          description: Defaults to trace started_at.
        ended_at:
          type: integer
          format: int64
          nullable: true
    IngestTrace:
      type: object
      properties:
        id:
          type: string
          nullable: true
          description: Auto-generated UUID if omitted. Max 128 characters.
        session_id:
          type: string
          nullable: true
        user_id:
          type: string
          nullable: true
        name:
          type: string
          default: ""
        status:
          type: string
          default: completed
        input:
          type: string
          nullable: true
          description: Subject to content storage policy.
        output:
          type: string
          nullable: true
          description: Subject to content storage policy.
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        started_at:
          type: integer
          format: int64
          nullable: true
          description: Defaults to server receive time.
        ended_at:
          type: integer
          format: int64
          nullable: true
        prompt_name:
          type: string
          nullable: true
        prompt_version:
          type: string
          nullable: true
        spans:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/IngestSpan"
...
    UpdateTrace:
      type: object
      properties:
        status:
          type: string
        output:
          type: string
        ended_at:
          type: integer
          format: int64
        input_tokens:
          type: integer
          format: int64
        output_tokens:
          type: integer
          format: int64
        cost:
          type: number
          format: double
          description: Cost in dollars. Converted to microdollars on storage.

    # -- LLM Tracing Query --
...
paths:
...
  /v1/traces:
    post:
      operationId: ingestTrace
      tags: [LLM Tracing Ingest]
      summary: Submit a single LLM trace
      description: |
        Ingests a trace with nested spans. Validates trace_id length (max 128)
        and span count (configurable max, default 100). If the SDK sends
        `cost=0` and a known model, the server calculates cost automatically
        using the built-in pricing table.
        Requires the `llm-tracing` feature flag.
      security:
        - hmac: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestTrace"
      responses:
        "200":
          description: Trace accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: accepted
        "400":
          description: Validation error (trace_id too long, too many spans).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or missing HMAC signature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/traces/{trace_id}:
    put:
      operationId: updateTrace
      tags: [LLM Tracing Ingest]
      summary: Update a running trace
      description: |
        Partially updates a trace (status, output, token counts, cost, end time).
        Useful for streaming traces that report results incrementally.
        Requires the `llm-tracing` feature flag.
      security:
        - hmac: []
      parameters:
        - name: trace_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrace"
      responses:
        "200":
          description: Trace updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
        "401":
          description: Invalid or missing HMAC signature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # =========================================================================
  # 8. LLM Tracing Query (feature: llm-tracing)
  # =========================================================================
...